requirements.txt
asgiref==3.8.1
boto3==1.34.150
botocore==1.34.150
certifi==2024.12.14
celery[redis]>=5.3
charset-normalizer==3.4.1
Django==5.0.6
django-celery-beat>=2.6
django-celery-results>=2.5
django-storages==1.14.4
ecdsa==0.19.0
idna==3.10
insightface>=0.7.3
jmespath==1.0.1
jose==1.0.0
numpy>=2.3.2
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
